unified OceanBC between HBModel and SWModel
Smallscale modifications to run on GPU

formatting

fix `test_windstress_short.jl` on CPU

use FreeSlip in GyreDriver.jl to match saved error

add some args, doesn't fix problem entirely

only types for dispatch

add Penetrable BC's for ConstantViscosity
