+++ src/Ocean/HydrostaticBoussinesq/LinearHBModel.jl
@@ -145,3 +145,16 @@ function wavespeed(lm::LinearHBModel, n⁻, _...)
     C = abs(SVector(lm.ocean.cʰ, lm.ocean.cʰ, lm.ocean.cᶻ)' * n⁻)
     return C
 end
+
+"""
+    boundary_state!(nf, ::LinearHBModel, args...)
+
+applies boundary conditions for the hyperbolic fluxes
+dispatches to a function in OceanBoundaryConditions.jl based on bytype defined by a problem such as SimpleBoxProblem.jl
+"""
+@inline function boundary_state!(nf, linear::LinearHBModel, args...)
+    ocean = linear.ocean
+    boundary_conditions = ocean.problem.boundary_conditions
+
+    return ocean_boundary_state!(nf, boundary_conditions, ocean, args...)
+end
